Advanced Backend Engineering/Restaurant-Menu System with Advanced Caching/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.1
redis==5.0.1
fastapi-cache2==0.2.1
psutil==5.9.6 